requirements.txt
anthropic>=0.30.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
jsonschema==4.23.0
python-dateutil==2.9.0.post0
httpx==0.27.2

pytest==8.3.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
